pyproject.toml
[project]
name = "bugeval"
version = "0.1.0"
description = "AI code review tools evaluation framework"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "anthropic>=0.40",
    "matplotlib>=3.8",
    "python-dotenv>=1.0",
    "aiohttp>=3.9",
    "google-genai>=1.0",
    "openai>=1.0",
]

[project.scripts]
bugeval = "bugeval.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.6",
    "pyright>=1.1",
    "flask>=3.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "standard"
venvPath = "."
venv = ".venv"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=bugeval --cov-report=term-missing"
